Fix minimap tile constants
